# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Notebook API",
        "description": "API endpoints for managing notebook entries",
        "version": "1.0.0"
    },
    "paths": {
        "/api/notebook": {
            "get": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Get a list of notebook entries",
                "description": "Get a list of notebook entries.",
                "operationId": "9f1b0487364942d1609f7381a2e225ed",
                "responses": {
                    "200": {
                        "description": "List of notebook entries",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/NotebookResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Create a new notebook entry",
                "description": "Create a new notebook entry.",
                "operationId": "0ba008f9429124e010071eedcd918bd1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "company": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "birthdate": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Notebook entry created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotebookResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/notebook/{id}": {
            "get": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Get a specific notebook entry by ID",
                "description": "Get a specific notebook entry by ID.",
                "operationId": "e1dc4e16a54d50b9eb3d75728ec33dc0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Notebook entry ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notebook entry details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotebookResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Notebook entry not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Update an existing notebook entry",
                "description": "Update an existing notebook entry.",
                "operationId": "4ef94d79cede0587746248b6d1c6484c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Notebook entry ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "company": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "birthdate": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notebook entry updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotebookResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Notebook entry not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Delete a notebook entry by ID",
                "description": "Delete a notebook entry by ID.",
                "operationId": "39c2b6ae51660dd4adb2c2a333db5f07",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Notebook entry ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notebook entry deleted successfully"
                    },
                    "404": {
                        "description": "Notebook entry not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "NotebookResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "full_name": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "birthdate": {
                        "type": "string"
                    },
                    "company": {
                        "type": "string"
                    },
                    "photo": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}